#45 카메라 줌 배율 개선 및 0.5배 광각 지원
- zoomLevel 타입을 Float으로 변경하여 소수점 단위 배율 지원
- 줌 버튼 클릭 시 배율 순환 로직 수정 (1.0x -> 2.0x -> 0.5x)
- UI 배율 표시를 위한 문자열 포맷팅 및 strings.xml 수정
- ImageProxy 사용 후 자원 해제 로직 추가로 메모리 관리 개선

+++ presentation/src/main/java/com/happyhouse/challa/presentation/camera/contract/CameraState.kt
@@ -20,7 +20,7 @@ data class CameraState(
     val lensFacing: CameraLensFacing = CameraLensFacing.BACK,
     val isFlashEnabled: Boolean = false,
     val captureRequest: PhotoCaptureRequest? = null,
-    val zoomLevel: Int = 1,
+    val zoomLevel: Float = 1f,
     val filterCount: Int = 8,
     val selectedFilterIndex: Int = 0,
     val rooms: ImmutableList<CameraRoomUiModel> = persistentListOf(),
